WIP: conditions cli, wasm tests, refactor

+++ plugin/condition/test/invalid-export/main.go
@@ -4,7 +4,7 @@ package main
 import "C"
 
 //export NotCheckCondition
-func NotCheckCondition(offerPtr, configPtr uint64) uint32 {
+func NotCheckCondition(offerPtr, configPtr uint64) uint64 {
 	return 0
 }
